g_trends v1/requirements_full.txt
# Full requirements for Keyword Trend Explorer - OpenAI Version
# This includes all packages currently installed in the development environment

altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.2.0
numpy==1.26.4
openai==1.102.0
packaging==25.0
pandas==2.2.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pycountry==24.6.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytrends==4.9.2
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.1.0
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
streamlit==1.48.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0